Issue #5044 - extend Jetty WS API Request/Response interfaces for jetty-websocket-server
