+++ jetty-websocket/websocket-jetty-client/src/main/java/org/eclipse/jetty/websocket/client/ClientUpgradeRequest.java
@@ -168,12 +168,6 @@ public final class ClientUpgradeRequest implements UpgradeRequest
     }
 
     @Override
-    public Object getSession()
-    {
-        throw new UnsupportedOperationException("HttpSession not available on Client request");
-    }
-
-    @Override
     public List<String> getSubProtocols()
     {
         return subProtocols;
