+++ jetty-websocket/websocket-jetty-client/src/main/java/org/eclipse/jetty/websocket/client/impl/DelegatedJettyClientUpgradeRequest.java
@@ -163,12 +163,6 @@ public class DelegatedJettyClientUpgradeRequest implements UpgradeRequest
     }
 
     @Override
-    public Object getSession()
-    {
-        return null;
-    }
-
-    @Override
     public Principal getUserPrincipal()
     {
         return null;
