+++ jetty-websocket/websocket-jetty-server/src/main/java/org/eclipse/jetty/websocket/server/JettyWebSocketServerContainer.java
@@ -41,6 +41,8 @@ import org.eclipse.jetty.websocket.core.WebSocketComponents;
 import org.eclipse.jetty.websocket.core.exception.WebSocketException;
 import org.eclipse.jetty.websocket.core.server.WebSocketServerComponents;
 import org.eclipse.jetty.websocket.server.config.JettyWebSocketServletContainerInitializer;
+import org.eclipse.jetty.websocket.server.internal.DelegatedServerUpgradeRequest;
+import org.eclipse.jetty.websocket.server.internal.DelegatedServerUpgradeResponse;
 import org.eclipse.jetty.websocket.server.internal.JettyServerFrameHandlerFactory;
 import org.eclipse.jetty.websocket.util.ReflectUtils;
 import org.eclipse.jetty.websocket.util.server.internal.FrameHandlerFactory;
@@ -87,7 +89,6 @@ public class JettyWebSocketServerContainer extends ContainerLifeCycle implements
     private static final Logger LOG = LoggerFactory.getLogger(JettyWebSocketServerContainer.class);
 
     private final WebSocketMapping webSocketMapping;
-    private final WebSocketComponents webSocketComponents;
     private final FrameHandlerFactory frameHandlerFactory;
     private final Executor executor;
     private final Configuration.ConfigurationCustomizer customizer = new Configuration.ConfigurationCustomizer();
@@ -105,7 +106,6 @@ public class JettyWebSocketServerContainer extends ContainerLifeCycle implements
     JettyWebSocketServerContainer(ServletContextHandler contextHandler, WebSocketMapping webSocketMapping, WebSocketComponents webSocketComponents, Executor executor)
     {
         this.webSocketMapping = webSocketMapping;
-        this.webSocketComponents = webSocketComponents;
         this.executor = executor;
 
         // Ensure there is a FrameHandlerFactory
@@ -129,7 +129,7 @@ public class JettyWebSocketServerContainer extends ContainerLifeCycle implements
             throw new WebSocketException("Duplicate WebSocket Mapping for PathSpec");
 
         webSocketMapping.addMapping(ps,
-            (req, resp) -> creator.createWebSocket(new JettyServerUpgradeRequest(req), new JettyServerUpgradeResponse(resp)),
+            (req, resp) -> creator.createWebSocket(new DelegatedServerUpgradeRequest(req), new DelegatedServerUpgradeResponse(resp)),
             frameHandlerFactory, customizer);
     }
